+++ pybamm/models/full_battery_models/base_battery_model.py
@@ -680,6 +680,13 @@ class BaseBatteryModel(pybamm.BaseModel):
                 raise pybamm.OptionError(
                     "x-average side reactions cannot be 'false' for SPM models"
                 )
+        if isinstance(self, pybamm.lithium_ion.SPM) and not isinstance(
+            self, pybamm.lithium_ion.MPM
+        ):
+            if options["particle size"] == "distribution":
+                raise pybamm.OptionError(
+                    "'particle size' should be 'single' for SPM and SPMe models"
+                )
         if isinstance(self, pybamm.lead_acid.BaseModel):
             if options["thermal"] != "isothermal" and options["dimensionality"] != 0:
                 raise pybamm.OptionError(