+++ pybamm/solvers/casadi_solver.py
@@ -261,13 +261,13 @@ class CasadiSolver(pybamm.BaseSolver):
                         dt_max = dt
                     count += 1
                     if count >= self.max_step_decrease_count:
+                        t_dim = t * model.timescale_eval
+                        dt_max_dim = dt_max * model.timescale_eval
                         raise pybamm.SolverError(
-                            "Maximum number of decreased steps occurred at t={}. Try "
-                            "solving the model up to this time only or reducing dt_max "
-                            "(currently, dt_max={})."
-                            "".format(
-                                t * model.timescale_eval, dt_max * model.timescale_eval
-                            )
+                            f"Maximum number of decreased steps occurred at t={t_dim}. "
+                            "Try solving the model up to this time only or reducing "
+                            f"dt_max (currently, dt_max={dt_max_dim}) and/or reducing "
+                            "the size of the time steps or period of the expeeriment."
                         )
                 # Check if the sign of an event changes, if so find an accurate
                 # termination point and exit
@@ -547,20 +547,11 @@ class CasadiSolver(pybamm.BaseSolver):
             rhs = model.casadi_rhs
             algebraic = model.casadi_algebraic
 
-            # When not in DEBUG mode (level=10), suppress warnings from CasADi
-            if (
-                pybamm.logger.getEffectiveLevel() == 10
-                or pybamm.settings.debug_mode is True
-            ):
-                show_eval_warnings = True
-            else:
-                show_eval_warnings = False
-
             options = {
+                "show_eval_warnings": False,
                 **self.extra_options_setup,
                 "reltol": self.rtol,
                 "abstol": self.atol,
-                "show_eval_warnings": show_eval_warnings,
             }
 
             # set up and solve